requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.22.0
plotly>=5.13.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.4.0
geopy>=2.3.0
scikit-learn>=1.2.0
joblib>=1.2.0 